requirements.txt
Flask
python-dateutil
adafruit-circuitpython-dht
RPi.GPIO
